% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/top-level.R
\name{roughness_wang}
\alias{roughness_wang}
\title{Wang et al.'s (2013) roughness model}
\usage{
roughness_wang(frequency_Hz, level_dB, detail = FALSE,
  include_phase_impact_factors = TRUE, msg = function(n, N, msg) if
  (interactive()) message(n, "/", N, ": ", msg))
}
\arguments{
\item{frequency_Hz}{(Numeric vector) Input frequencies (Hz).}

\item{level_dB}{(Numeric vector) Input sound levels (dB), of the same length as
\code{frequency_Hz}.}

\item{detail}{(Logical scalar) Whether to return detailed output information.}

\item{include_phase_impact_factors}{(Logical scalar)
Whether to include phase impact factors in roughness computation.
Set to \code{TRUE} to recover the original specifications of Wang et al. (2013).
However, disabling this feature (by setting the parameter to \code{FALSE})
seems to result in better estimation of perceptual consonance.}

\item{msg}{Function to be called to give progress updates.
This function should accept three arguments:
\code{n}, an integer identifying the current position in the pipeline,
\code{N}, an integer identifying the length of the pipeline,
and \code{msg}, a string providing a longer-format description
of the current position in the pipeline.}
}
\value{
If \code{detail == FALSE}, a numeric vector of roughnesses,
otherwise a list containing detailed algorithm output.
}
\description{
Gets the roughness of an acoustic spectrum according to the model of Wang et al. (2013).
}
\note{
This implementation is designed for sparse input spectra, that is,
when \code{frequency_Hz} and \code{level_dB} are both relatively
short vectors.
}
\references{
\insertRef{Wang2013}{wang13}
}
